Rename to update_settings_container.dart
