+++ app_update/lib/src/parser/models/update_config_model.dart
@@ -1,4 +1,4 @@
-import '../../shared/update_status_wrapper.dart';
+import '../../shared/update_settings_container.dart';
 import 'release_config.dart';
 import 'source_config.dart';
 import 'versions_settings_config.dart';
